improved aiming

+++ SnowFight/Assets/Weapons/Projectile.js
@@ -1,4 +1,4 @@
 var dmg : int = 5;
 var speed : int = 30;
-var reloadTime : int = 30;
+var reloadTime : float = 30.0;
 var snowCosts :int = 0;
